requirements.txt
# ============================================================================
# AutoTrader — Python Dependencies
# ============================================================================
# Install: pip install -r requirements.txt
# Python version: 3.11+
# ============================================================================

# Telegram client (user account API for listening to channels)
Telethon==1.37.0

# MT5 bridge for Linux (communicates with MT5 running in Wine)
mt5linux==0.1.9

# Google Gemini AI SDK (primary signal parser)
google-generativeai==0.8.5

# Groq AI SDK (failover signal parser)
groq==0.15.0

# Configuration loading from .env files
python-dotenv==1.0.1

# Data validation and serialization
pydantic==2.10.5
pydantic-settings==2.7.1

# Async HTTP server (health check endpoint)
aiohttp==3.11.11

# Async SQLite (signal history, trade log, deduplication)
aiosqlite==0.20.0

# Telegram Bot API (notification alerts to admin)
python-telegram-bot==21.10

# HTTP client for Groq API calls
httpx==0.28.1

# Cryptographic hashing (stdlib, listed for clarity)
# hashlib — built-in, no install needed

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
